requirements.txt
Flask
gunicorn
python-dotenv
supabase